requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
flake8==7.1.1
Markdown==3.7
mccabe==0.7.0
psycopg2==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
sqlparse==0.5.2
drf-spectacular
pillow